resume_jd_matcher/requirements.txt
flask
PyMuPDF
python-dotenv
google-generativeai